refactor: decompose scheduler/index.ts into focused modules
Split the 631-line startScheduler() monolith into seven focused modules:
- state.ts: shared mutable state container for late-binding closures
- validation.ts: agent discovery, config validation, credential checks
- runner-setup.ts: scale enforcement + runner pool creation
- call-dispatcher.ts: al-call dispatch wiring
- cron-setup.ts: cron jobs, enable/disable handlers, initial runs
- gateway-setup.ts: gateway startup with late-binding controlDeps
- shutdown.ts: graceful SIGINT/SIGTERM shutdown

Also extracted registerWebhookBindings() into webhook-setup.ts as a
shared function used by both initial setup and hot-reload (watcher.ts),
eliminating duplicated webhook wiring logic.

The orchestrator (index.ts) is now ~170 lines. No behavior changes.
